Generic function to access data from functions in Atlas database

Return data objects stored obtained through endpoints corresponding to the name of the database function.

Usage

db_call_function(
  name,
  schema = "public",
  output_geometry = FALSE,
  output_flatten = TRUE,
  .token = NULL,
  ...
)

Arguments

name

character. Name of the atlas function to be accessed. data. Can be either api, public or atlas_api.

schema

Optional. character Schema from the database where is located the data. Can be either api, public or atlas_api.

output_geometry

Optional. logical default FALSE. If TRUE, returns an sf object using the geometry column from the function output.

output_flatten

Optional. logical default FALSE. If TRUE, returns a data.frame object with nested objects flattened.

.token

Optional. character Bearer token providing access to the web api.

...

character or numeric scalar or vector. Arguments to be passed to the function.

Value

tibble with rows associated with Atlas data object

Details

This function is designed to interface with a web API deployed with PostgREST