Generic function to post data into Atlas databases

Data records attributes (list names or columns) must corresponds to the table columns it is sent to. (ie. taxa, observations, datasets, etc).

Usage

db_write_table(
  table_name,
  data,
  ...,
  schema = "public",
  host = ATLAS_API_V4_HOST(),
  .page_limit = 50000,
  .token = ATLAS_API_TOKEN(),
  .cores = 4
)

Arguments

table_name

character. Name of the atlas data object destination. Corresponds to the name of a within Atlas Postgresql database, stored within schema public ou api

data

list or data.frame. Either a single record as a list with names as attributes or multiple records as a data.frame with attributes as columns with attibutes corresponding to the destination column tables.

...

character or numeric scalar or vector. Additional parameters to provide to the request. May correspond to postgrest http request parameters and syntax.

schema

character Schema from the database where is located the data object is located. Accept either values api or public (default)

.page_limit

integer Count of objects returned through pagination

.token

character Bearer token providing access to the web api

.cores

integer default 4. Number of cores used to parallelize and improve rapidity

Details

This function is designed to interface with a web API deployed with PostgREST