Wrapper function of kableExtra::row_spec

Apply some default formatting to the header row of a kable table. Should be called after any calls to kableExtra::column_spec().

fmt_kbl_header(
  kable_input,
  row = 0,
  align = "c",
  extra_css = "border-bottom: 0.16em solid #111111",
  ...
)

Arguments

kable_input

Output of knitr::kable() with format specified

row

A numeric value or vector indicating which row(s) to be selected. You don't need to count in header rows or group labeling rows.

align

A character string for cell alignment. For HTML, possible values could be l, c, r plus left, center, right, justify, initial and inherit while for LaTeX, you can only choose from l, c & r.

extra_css

Extra css text to be passed into the cells of the row. Note that it's not for the whole row.

...

Additional arguments passed to kableExtra::row_spec()

Value

A kable object.

Examples

fmt_kbl(mtcars[, 1:3], align = c("r", "c", "r"),
    col.names = c("Column 1", "Column 2", "Column 3"),
    caption = "Example Table Title") |>
  kableExtra::column_spec(1, width = "20em") |>
  fmt_kbl_header()
#> <table class="table lightable-classic" style=' font-family: "Arial Narrow", "Source Sans Pro", sans-serif; margin-left: auto; margin-right: auto;'>
#> <caption>Example Table Title</caption>
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;text-align: center;border-bottom: 0.16em solid #111111">   </th>
#>    <th style="text-align:right;text-align: center;border-bottom: 0.16em solid #111111"> Column 1 </th>
#>    <th style="text-align:center;text-align: center;border-bottom: 0.16em solid #111111"> Column 2 </th>
#>    <th style="text-align:right;text-align: center;border-bottom: 0.16em solid #111111"> Column 3 </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Mazda RX4 </td>
#>    <td style="text-align:right;"> 21.0 </td>
#>    <td style="text-align:center;"> 6 </td>
#>    <td style="text-align:right;"> 160.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Mazda RX4 Wag </td>
#>    <td style="text-align:right;"> 21.0 </td>
#>    <td style="text-align:center;"> 6 </td>
#>    <td style="text-align:right;"> 160.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Datsun 710 </td>
#>    <td style="text-align:right;"> 22.8 </td>
#>    <td style="text-align:center;"> 4 </td>
#>    <td style="text-align:right;"> 108.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Hornet 4 Drive </td>
#>    <td style="text-align:right;"> 21.4 </td>
#>    <td style="text-align:center;"> 6 </td>
#>    <td style="text-align:right;"> 258.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Hornet Sportabout </td>
#>    <td style="text-align:right;"> 18.7 </td>
#>    <td style="text-align:center;"> 8 </td>
#>    <td style="text-align:right;"> 360.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Valiant </td>
#>    <td style="text-align:right;"> 18.1 </td>
#>    <td style="text-align:center;"> 6 </td>
#>    <td style="text-align:right;"> 225.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Duster 360 </td>
#>    <td style="text-align:right;"> 14.3 </td>
#>    <td style="text-align:center;"> 8 </td>
#>    <td style="text-align:right;"> 360.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Merc 240D </td>
#>    <td style="text-align:right;"> 24.4 </td>
#>    <td style="text-align:center;"> 4 </td>
#>    <td style="text-align:right;"> 146.7 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Merc 230 </td>
#>    <td style="text-align:right;"> 22.8 </td>
#>    <td style="text-align:center;"> 4 </td>
#>    <td style="text-align:right;"> 140.8 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Merc 280 </td>
#>    <td style="text-align:right;"> 19.2 </td>
#>    <td style="text-align:center;"> 6 </td>
#>    <td style="text-align:right;"> 167.6 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Merc 280C </td>
#>    <td style="text-align:right;"> 17.8 </td>
#>    <td style="text-align:center;"> 6 </td>
#>    <td style="text-align:right;"> 167.6 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Merc 450SE </td>
#>    <td style="text-align:right;"> 16.4 </td>
#>    <td style="text-align:center;"> 8 </td>
#>    <td style="text-align:right;"> 275.8 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Merc 450SL </td>
#>    <td style="text-align:right;"> 17.3 </td>
#>    <td style="text-align:center;"> 8 </td>
#>    <td style="text-align:right;"> 275.8 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Merc 450SLC </td>
#>    <td style="text-align:right;"> 15.2 </td>
#>    <td style="text-align:center;"> 8 </td>
#>    <td style="text-align:right;"> 275.8 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Cadillac Fleetwood </td>
#>    <td style="text-align:right;"> 10.4 </td>
#>    <td style="text-align:center;"> 8 </td>
#>    <td style="text-align:right;"> 472.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Lincoln Continental </td>
#>    <td style="text-align:right;"> 10.4 </td>
#>    <td style="text-align:center;"> 8 </td>
#>    <td style="text-align:right;"> 460.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Chrysler Imperial </td>
#>    <td style="text-align:right;"> 14.7 </td>
#>    <td style="text-align:center;"> 8 </td>
#>    <td style="text-align:right;"> 440.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Fiat 128 </td>
#>    <td style="text-align:right;"> 32.4 </td>
#>    <td style="text-align:center;"> 4 </td>
#>    <td style="text-align:right;"> 78.7 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Honda Civic </td>
#>    <td style="text-align:right;"> 30.4 </td>
#>    <td style="text-align:center;"> 4 </td>
#>    <td style="text-align:right;"> 75.7 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Toyota Corolla </td>
#>    <td style="text-align:right;"> 33.9 </td>
#>    <td style="text-align:center;"> 4 </td>
#>    <td style="text-align:right;"> 71.1 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Toyota Corona </td>
#>    <td style="text-align:right;"> 21.5 </td>
#>    <td style="text-align:center;"> 4 </td>
#>    <td style="text-align:right;"> 120.1 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Dodge Challenger </td>
#>    <td style="text-align:right;"> 15.5 </td>
#>    <td style="text-align:center;"> 8 </td>
#>    <td style="text-align:right;"> 318.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> AMC Javelin </td>
#>    <td style="text-align:right;"> 15.2 </td>
#>    <td style="text-align:center;"> 8 </td>
#>    <td style="text-align:right;"> 304.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Camaro Z28 </td>
#>    <td style="text-align:right;"> 13.3 </td>
#>    <td style="text-align:center;"> 8 </td>
#>    <td style="text-align:right;"> 350.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Pontiac Firebird </td>
#>    <td style="text-align:right;"> 19.2 </td>
#>    <td style="text-align:center;"> 8 </td>
#>    <td style="text-align:right;"> 400.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Fiat X1-9 </td>
#>    <td style="text-align:right;"> 27.3 </td>
#>    <td style="text-align:center;"> 4 </td>
#>    <td style="text-align:right;"> 79.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Porsche 914-2 </td>
#>    <td style="text-align:right;"> 26.0 </td>
#>    <td style="text-align:center;"> 4 </td>
#>    <td style="text-align:right;"> 120.3 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Lotus Europa </td>
#>    <td style="text-align:right;"> 30.4 </td>
#>    <td style="text-align:center;"> 4 </td>
#>    <td style="text-align:right;"> 95.1 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Ford Pantera L </td>
#>    <td style="text-align:right;"> 15.8 </td>
#>    <td style="text-align:center;"> 8 </td>
#>    <td style="text-align:right;"> 351.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Ferrari Dino </td>
#>    <td style="text-align:right;"> 19.7 </td>
#>    <td style="text-align:center;"> 6 </td>
#>    <td style="text-align:right;"> 145.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Maserati Bora </td>
#>    <td style="text-align:right;"> 15.0 </td>
#>    <td style="text-align:center;"> 8 </td>
#>    <td style="text-align:right;"> 301.0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 20em; "> Volvo 142E </td>
#>    <td style="text-align:right;"> 21.4 </td>
#>    <td style="text-align:center;"> 4 </td>
#>    <td style="text-align:right;"> 121.0 </td>
#>   </tr>
#> </tbody>
#> </table>