Center and Decimal Align Tables

Automatic formatting for tables that should "just work" for most use cases. For more fine-grained control, see wjake::fmt_digits() and padding.

fmt_table(
  df,
  dec_dig = 1,
  prop_dig = 3,
  corr_dig = 3,
  output = NULL,
  keep_boundary = FALSE,
  ...
)

Arguments

df

A data frame or tibble to be formatted for printing in output.

dec_dig

The number of decimal places to include for numbers, e.g., dec_dig = 1 for 16.5.

prop_dig

The number of decimal places to include for numbers bounded between [0,1], e.g., prop_dig = 2 for .35.

corr_dig

The number of decimal places to include for numbers bounded between [-1,1], e.g., corr_dig = 3 for .205.

output

The output format of the table. One of "latex" or "html". Automatically pulled from document output type if not specified.

keep_boundary

Whether to preserve true values of boundaries (i.e., min_value and max_value). For example if digits = 3, min_value = 0, and keep_boundary = FALSE, a value exactly equal to 0 will become "<.001". If keep_boundary = TRUE, then a value of 0 will remain "0.000", and other small values (e.g., 0.000001) will continue to be replaced with "<.001".

...

Additional arguments passed to wjake::fmt_digits().

Value

A tibble with the same rows and columns as df, with numbers formatted consistently and padded for alignment when printed.

See also

Examples

pcts <- data.frame(n = 0:5, p = 0.5 * (0:5))
fmt_table(pcts)
#>   n   p
#> 1 0 0.0
#> 2 1 0.5
#> 3 2 1.0
#> 4 3 1.5
#> 5 4 2.0
#> 6 5 2.5