Get datasets

The function downloads and returns a list or dataframe containing a description for each requested dataset where a row corresponds to an record. The function returns a dataframe for all datasets if no parameters are specified. The function filters returned datasets record by attributes corresponding to atlas table columns specified as parameters (ie. id, title, original_source, etc) with accepted values either being scalar or vector for single or multiple records

Usage

get_datasets(id = NULL, ...)

Arguments

id

Optional. integer scalar or vector. Returns a dataframe for the dataset with the specified id

...

Optional. scalar or vector. Returns a dataframe filtered by the atlas datasets table columns specified as parameter

Value

tibble with rows associated with Atlas data object

Examples

# Returns all available datasets records in atlas
datasets <- get_datasets()

# Returns all datasets filtered by the column `original_source`
datasets <- get_datasets(original_source = "eBird")

# Returns all datasets filtered by the column id corresponding to eBird
datasets <- get_datasets(id = 55:102)