Generic function to access data from Atlas databases

Returns Atlas data objects stored within the database. Queried data objects are obtained through endpoints corresponding to the name of the database object storing the required data (ie. taxa, observations, datasets, etc).

get_gen(
  endpoint,
  ...,
  .page_limit = 5e+05,
  .schema = "public",
  .cores = 4,
  .page_parameters = DEFAULT_PAGE_PARAMETERS,
  .n_pages = NA
)

Arguments

endpoint
character. Name of the atlas data object to be accessed. Corresponds to the name of a table, view or stored procedure within Atlas Postgresql database, stored within schema public ou api

...
character or numeric scalar or vector. Additional parameters to filter the returned objects. Accepts parameters corresponding to an object attribute or table columns (ie. id = c(3,4)) or a specific parameters related to postgrest http request parameters and syntax(select = ...)

.page_limit
integer Count of objects returned through pagination

.schema
character Schema from the database where is located the data object is located. Accept either values api or public (default)

.cores
integer default 4. Number of cores used to parallelize and improve rapidity

.page_parameters
list default list(limit = "limit", offset = "offset"). Parameters names sent to http query to be consumed by the endpoint for pagination

.token
character Bearer token providing access to the web api

..n_pages
integer default NULL. Estimated if none provided

Value

tibble with rows associated with Atlas data object

Details

This function is designed to interface with a web API deployed with PostgREST

Examples

# Returns all available taxa records in atlas
taxa <- get_gen("taxa")

# Returns all taxa filtered by the column id values
taxa <- get_gen("taxa", id = c(188, 201, 294, 392))

# Returns first 100 sampling_points from schema `api`
sampling_points <- get_gen(
  "bird_sampling_points",
  .schema = "api",
  limit = 100
)

# Returns first 100 observations with selected columns and joined columns
get_gen("observations",
  limit = 100,
  select = "geom,year_obs,month_obs,day_obs,obs_value,taxa(scientific_name)"
)