Get observations

The function downloads and returns a list or dataframe containing observations with a single record per row. The function returns a dataframe containing all records if no parameters are specified (STRONGLY DISCOURAGED DUE TO LARGE AMOUNT OF OBSERVATIONS CONTAINED WITHIN ATLAS). The function filters returned records by attributes corresponding to atlas table columns specified as parameters (ie. id, year_obs, id_taxa, id_datasets, id_variables, fid_region, etc) with accepted values either being scalar or vector for single or multiple records.

Usage

get_observations(
  id = NULL,
  year = NULL,
  id_taxa = NULL,
  region_fid = NULL,
  geometry = TRUE,
  within_quebec = NULL,
  ...
)

Arguments

id

Optional. integer scalar or vector. Returns a dataframe for the observation with the specified id

year

Optional. integer scalar or vector. Returns a dataframe for the observations related to the id_taxa

id_taxa

Optional. integer scalar or vector. Returns a dataframe for the taxa_obs record related to the value. id_taxa is translated to id_taxa_obs

within_quebec

Optional. logical default NULL. If TRUE, returns only observations within Quebec. If FALSE, returns only observations outside Quebec. If NULL, returns all observations.

...

Optional. scalar or vector. Returns a dataframe filtered by the atlas observations table columns specified as parameter (ie. id_datasets, id_variables, month_obs, day_obs)

fid_region

Optional. integer scalar or vector. Returns observations for the region with the specified id

.cores

Optional. integer default 4. Number of cores used to parallelize and improve rapidity

Value

tibble with rows associated with Atlas observations