Get regions

The function downloads and returns a list or dataframe containing geometries and attributes for each requested region where a row corresponds to a region. Regions are organized by type (ie. Ecological Region, Administrative Region, and Hexagonal Grfid) and each region type is organized by scale. It is recommended to explore regions using the Biodiversité Québec Atlas web application to determine the region type and scale of interest.

Usage

get_regions(fid = NULL, type = NULL, scale = NULL, geometry = TRUE, ...)

Arguments

type

Optional. character scalar or vector. Returns a dataframe filtered by the atlas regions table column type specified as parameter. Accepted values are cadre_eco, admin, and hex

...

Optional. scalar or vector. Returns a dataframe filtered by the atlas regions table columns specified as parameter.

fid_region

Optional. integer scalar or vector. Returns a dataframe for the region with the specified fid

zoom

Optional. integer scalar or vector. Returns a dataframe

Details

IMPORTANT : Running the function without any parameters will return all regions in the atlas. This can be a very large dataset and may take a long time to download. It is recommended to use the fid parameter to return a specific region or the type parameter to return a specific type of region. It is also possible to exclude the geometry from the results by setting the geometry parameter to FALSE.