Save a ggplot2 graphic

This is a wrapper around ggplot2::ggsave() with some ATLAS-specific defaults. The aspect ratio is fixed to 0.618 (the golden ratio) unless the height is manually defined. Plots are automatically spell checked and warnings are returned if there are possible mistakes. Finally, plots saved as a pdf have the fonts embedded using extrafont::embed_fonts().

ggsave2(
  plot = ggplot2::last_plot(),
  filename,
  device = NULL,
  path = NULL,
  width = 8,
  height = NULL,
  units = "in",
  dir = c("h", "v"),
  dpi = "retina",
  ...
)

Arguments

plot
Plot to save, defaults to last plot displayed.

filename
File name to create on disk.

device
Device to use. Can either be a device function (e.g. png), or one of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).

path
Path of the directory to save plot to: path and filename are combined to create the fully qualified file name. Defaults to the working directory.

width
Plot size in units ("in", "cm", or "mm").

height
Plot size in units ("in", "cm", or "mm"). If not supplied, uses 0.618 * width when dir = "h" and 1.618 * width when dir = "v".

units
Plot size in units ("in", "cm", "mm", or "px"). If not supplied, uses the size of current graphics device.

dir
Orientation of the plot. One of h (default) for horizontal or v for vertical.

dpi
Plot resolution. Also accepts a string input: "retina" (320), "print" (300), or "screen" (72). Applies only to raster output types.

...
Additional arguments passed to ggplot2::ggsave()