Function to insert datasets in Atlas. Formatting should follow https://eml.ecoinformatics.org/schema/

This function is designed to interface with a web API deployed with PostgREST Function is not exported, as it is meant for internal use only. It can be called like so: ratlas:::post_dataset()

Usage

post_dataset(
  title,
  creator,
  pub_date = NULL,
  abstract = NULL,
  keyword_set = NULL,
  distribution = NULL,
  alternate_identifier = NULL,
  licensed = NULL,
  associated_party = NULL,
  contact,
  metadata_provider = NULL,
  additional_info = NULL,
  language = NULL,
  publisher = NULL,
  coverage = NULL,
  methods = NULL,
  additional_metadata = NULL,
  data_type,
  source_eml_url = NULL,
  source_alias,
  shareable_data = FALSE,
  schema = "api"
)

Arguments

title

character. Name of the dataset.

creator

list. List of responsibleParty objects created with EML::set_responsibleParty().

pub_date

Optional. character. Publication date of the dataset in "YYYY-MM-DD" format.

abstract

Optional. character. Abstract of the dataset.

keyword_set

Optional. list. List of keywordSet objects created manually.

distribution

Optional. list. List of list distribution objects created manually.

alternate_identifier

Optional. list. List of alternateIdentifier character. DOI is declared here.

licensed

Optional. list. List of license informations created manually.

associated_party

Optional. list. List of responsibleParty objects created manually.

contact

list. List of responsibleParty objects created with EML::set_responsibleParty().

metadata_provider

Optional. list. List of responsibleParty objects created with EML::set_responsibleParty().

additional_info

Optional. list. List of additional information about the dataset created with EML::set_TextType() to a markdown tag.

language

Optional. character. Language of the dataset.

publisher

Optional. list. List of responsibleParty objects created with EML::set_responsibleParty().

coverage

Optional. list. List of coverage objects created with EML::set_coverage().

methods

Optional. jsonb. Methods section created with EML::set_methods() or manually.

additional_metadata

Optional. character. Citation string, nested in metadata, citation.

Non-EML parameters:

data_type

character. Type of data, example: occurrence, time series, etc.

source_eml_url

Optional. character. URL of the source EML.

source_alias

character. Internal alias to recognize the dataset.

shareable_data

boolean. Default FALSE, if TRUE it means we have the license to share the data.

schema

obligatory api, as function resides in this schema on the backend.

Value

newly generated dataset UUID.

Details

EML parameters: