Append row and/or column summaries

Add row and/or column summaries (e.g., total counts) to a data frame.

append_summary(df, ..., row = TRUE, col = TRUE, .f = sum, args = NULL)

Arguments

df

A data frame to append summaries to.

...

Unquoted names of columns to be included in the summary.

row

logical indicating whether a summary row should be added (i.e., summarizing each column).

col

logical indicating whether a summary column should be added (i.e., summarizing each row).

.f

Function to use for calculating summaries.

args

A named list of additional arguments to pass to .f.

Value

A data frame with the summary row and/or column appended

Examples

df <- head(penguins[, c("species", "bill_len", "flipper_len")])
append_summary(df, bill_len, flipper_len, row = TRUE, col = TRUE, .f = sum)
#>   species bill_len flipper_len   sum
#> 1  Adelie     39.1         181 220.1
#> 2  Adelie     39.5         186 225.5
#> 3  Adelie     40.3         195 235.3
#> 4  Adelie       NA          NA    NA
#> 5  Adelie     36.7         193 229.7
#> 6  Adelie     39.3         190 229.3
#> 7    <NA>       NA          NA    NA

append_summary(df, where(is.numeric), .f = sum, args = list(na.rm = TRUE))
#>   species bill_len flipper_len    sum
#> 1  Adelie     39.1         181  220.1
#> 2  Adelie     39.5         186  225.5
#> 3  Adelie     40.3         195  235.3
#> 4  Adelie       NA          NA    0.0
#> 5  Adelie     36.7         193  229.7
#> 6  Adelie     39.3         190  229.3
#> 7    <NA>    194.9         945 1139.9

append_summary(df, bill_len, flipper_len, row = FALSE, .f = mean)
#>   species bill_len flipper_len   mean
#> 1  Adelie     39.1         181 110.05
#> 2  Adelie     39.5         186 112.75
#> 3  Adelie     40.3         195 117.65
#> 4  Adelie       NA          NA     NA
#> 5  Adelie     36.7         193 114.85
#> 6  Adelie     39.3         190 114.65